LayoutTests/editing/selection/ios/caret-when-focusing-editable-element-with-selection-expected.txt

This test verifies that when setting the selection inside an editable element while the web view is not first responder, the caret should appear after tapping the editable element to show the keyboard. This test requires WebKitTestRunner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS caretRect.left is 9
PASS caretRect.top is 21
PASS caretRect.width is 2
PASS caretRect.height is 24
PASS successfullyParsed is true

TEST COMPLETE
